Refactor database init to remove lazy cyclic require
